+++ web/src/app/trade/trade.component.ts
@@ -1,6 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
 
+import { BillComponent } from '../bill/bill.component';
+import { ConfirmModel, ConfirmModelComponent } from '../components/confirm-model/confirm-model.component';
 import { Trade } from '../model/trade.model';
 import { ToastService } from '../service/toast-service';
 import { TradeService } from '../service/trade-service';
@@ -16,7 +18,7 @@ export class TradeComponent implements OnInit {
   constructor(
     private service: TradeService,
     private toastService: ToastService,
-    private modalService: NgbModal
+    private modalService: NgbModal,
   ) { }
 
   ngOnInit(): void {
@@ -41,7 +43,7 @@ export class TradeComponent implements OnInit {
       item.transactions = transactions;
       let modelRef = this.modalService.open(AddTradeComponent, { size: 'lg' });
       modelRef.componentInstance.isEditMode = true;
-      modelRef.componentInstance.inventory = item;
+      modelRef.componentInstance.trade = item;
       modelRef.result.then(() => {
         this.toastService.info("Dealing updated.");
         this.fetchList();
@@ -50,9 +52,32 @@ export class TradeComponent implements OnInit {
   }
 
   delete(item: Trade) {
-    this.service.Delete(item).subscribe(() => {
-      this.toastService.info("Dealing deleted successfully");
-      this.fetchList();
+    let modelRef = this.modalService.open(ConfirmModelComponent);
+    const data: ConfirmModel = {
+      title: "Remove Trade entry of" + item.Dealer_Name,
+      description: "Are you sure you want to remove Trade entry of customer " + item.Dealer_Name + " ?",
+    };
+    modelRef.componentInstance.data = data;
+    modelRef.result.then((resp) => {
+      if (resp) {
+        this.service.Delete(item).subscribe(() => {
+          this.toastService.info("Dealing deleted successfully");
+          this.fetchList();
+        });
+      }
+    });
+
+  }
+
+  print(item: Trade) {
+    this.service.GetTransactions(item.trade_id).subscribe((transactions) => {
+      item.transactions = transactions;
+      let modelRef = this.modalService.open(BillComponent, { size: 'lg' });
+      modelRef.componentInstance.isEditMode = true;
+      modelRef.componentInstance.trade = item;
+      modelRef.result.then(() => {
+      });
     });
   }
+
 }
